Add FPDF_StructElement_GetMarkedContentIdCount() and FPDF_StructElement_GetMarkedContentIdAtIndex()
FPDF_StructElement_GetMarkedContentID() provides access to the MCID in a
struct element. It assumes /K in a struct element dictionary is a number.
This may be a case sometimes, but /K can also be an array with mixed
types. See ISO 32000-1:2008, section 14.7.6 for example.

This change supersedes FPDF_StructElement_GetMarkedContentID(). It
provides a mechanism to access MCID at an index while supporting the
case where struct element dictionary is a number. After this change,
FPDF_StructElement_GetMarkedContentID() becomes redundant and should be
deprecated.

Change-Id: Icecd39fe3ea788e7f85851c2b7eb89453d083b2f
Bug: pdfium:1773