[Skia] Fix the issue that texts are rendered narrower than expected.
When Skia/SkiaPaths is enabled, render a substituted glyph using its
path instead of using CFX_SkiaDeviceDriver when its width defined in
the PDF is narrower than the default glyph width.

This CL also update the expectations of the embedder tests, pixel tests
and corpus tests which are affected by this change.

Bug: pdfium:1752
Change-Id: If0f990fb4eaeda509a74b33d8086e4b8e69cf074