Simplify CPDF_InteractiveForm::AddNativeInteractiveFormFont().
Avoid having a parameter that passes a pointer by reference, which can
be confusing. Do the same for internal methods like AddFont() and
InitDict(). Then it becomes obvious that InitDict() always receives a
newly-created dictionary, and some of the checks it does are always true
or always false. Remove those checks to simplfy the code.

Change-Id: Idf81695e97b5b0102a07d1dfb9934a98d8db4b01