Add a test for rendering embedded font with narrow/wide glyph width.
Pixel test bug_845697.in usually gets rendered with FoxitSansMM as the
substitute font. To observe how glyphs are scaled when their assigned
widths change in the PDF, this CL adds a new test with NotoSansRegular
font embedded, and test on rendering behaviors when assigned character
width is either narrower or wider than the glyph's default width.

Bug: pdfium:1752
Change-Id: I5bd191c143addda6f450dff062765122eafb3d26