BugFix: Fix TypeError on executing encode_pdf_filter.py by making property getters classmethods
This CL is a follow-up of I205d254514f1ae667af818c789b5e6d89a061c68, to
fix the a bug introduced by that commit. That commit made name and aliases
properties methods. However, these properties are accessed by
classmethods, so the callers get a Property object rather than a value.
This CL fixes this issue my marking the propeties as classmethods as
well.

Bug: pdfium:1776
Change-Id: Ic668f920450d23a0cdca173c909f427c7d63f17c