Improve TrueType font fallback check for glyph ID 0.
Change the heuristics used to check for font fallback when drawing
TreuType fonts with glyph ID 0. In some cases, 0 means use glyph 0, and
in other cases, it means the glyph is invalid. Right now, the code tries
to use a fallback font, and if that fails, go back to the original font.
Instead of doing this back-and-forth, look at the heuristics and decide
one way or the other.

Bug: chromium:1271578
Change-Id: I71a0a6607162602d8e3fba07e80546ed3a1ce548