Make sure some array accesses in CFX_FontMapper never go out of bounds.
Existing callers to GetFaceName() and RawBytesForIndex() never pass in
an out of bound index value. Nevertheless, use CHECK_LT() to ensure to
make sure it stays that way.

Also remove an existing check in RawBytesForIndex() for a null
`m_pFontInfo`, which can never be true because that would imply there
are no font faces. In which case, RawBytesForIndex() should never be
called.

Change-Id: I4b48bf4881308ca31653e8579f544c7c6a30e716