Fix FT_UnicodeFromCharCode() encoding mismatch.
Now that enum class FontEncoding exists, it is obvious that
FT_UnicodeFromCharCode() is not taking in the right encoding value in
some cases. Looking at the existing callers to FT_UnicodeFromCharCode()
and its behavior, the call only works as intended when the encoding is
`FT_ENCODING_APPLE_ROMAN`. Thus simplify FT_UnicodeFromCharCode() to
UnicodeFromAppleRomanCharCode(). Remove the FT_UnicodeFromCharCode()
calls that do not make sense, where FT_UnicodeFromCharCode() always
returns 0, and simplify the caller code.

Change-Id: I948886d2ff90fc7ffd4ac14eb9309cf34ba29a49