Add a test for rendering image object with the correct color space.
Add a pixel test based on the attachments in crbug.com/pdfium/1723
which represents the issue that PDFium should not render the PDF with
the color space in /Resource dictionary instead of the the color space
specified in the image object.

Once the bug is fixed, the rendering result should match the expectation.

Bug: pdfium:1723
Change-Id: I2564a504c00afb3efc7fbb297db7e35382b201ca