[Skia] Fix mask blending issues.
When painting a mask-format image, it can be treated as an individual
image ready to be painted or it can be used as the mask of another
image. For the former case, it's used as an image instead of a color
filter.

This CL applies the following changes for these two cases:
- Adds an extra parameter `merge` to SetBitmapPaint() to make sure we
  only apply color filter when merging a mask onto another image object.
- Sets paint color when simply painting a mask format image object since
  a image format bitmap must be 1-bit-per-component and doesn't contain
  the color information from the graphic state.

Bug: pdfium:1746, pdfium:1747, pdfium:1749
Change-Id: I6f363d0316375eea302a29804def049934851ddc