[Skia] Fix mask blending issues.
When painting a mask-format image, it can be treated as an individual
image ready to be painted or it can be used as the mask of another
image. For the former case, it's used as an image instead of a color
filter.

Based on the existing SetBitmapPaint(), this CL applies the following
changes for these two cases:

- Call SetBitmapPaint() only for the former case: Make this function set
  paint color since a image format bitmap must be 1-bit-per-component
  and doesn't contain the color information from the graphic state.
  Also avoids setting color filter in this function.

- Creates SetBitmapPaintForMerge() which applies color filter to the
  SkPaint and only call this function for the latter case.

Bug: pdfium:1746, pdfium:1747, pdfium:1749
Change-Id: I6f363d0316375eea302a29804def049934851ddc