[Skia/SkiaPaths] Fix text rendering when font data is missing.
CFX_SkiaDeviceDriver relies on font data to build SkTextBlob for drawing
text. When the font data is missing, these blobs will have the bounds of
zero area and eventually be rejected by Skia for drawing.

This CL makes text rendering fall back on drawing glyph bitmaps when the
font data is missing.

Bug: pdfium:1354, pdfium:1743
Change-Id: Iaedaab137671399955bffc04dc29c23571c73b2f