Avoid floating-point loop counter in vcgen_dash::calc_dash_start().
When a floating-point loop counter is incremented/decreased by an amount
that is too small to change its value given its precision, the loop will
get stuck and cause the process to hang.

This CL changes vcgen_dash::calc_dash_start() to use an integer loop
counter instead of a floating-point one with the precision of 3 decimal
digits.

Bug: pdfium:1822
Change-Id: I3750f8f59260df2c06fd734edce4d57535db8e12