Fix component build when Skia is enabled
SkFontMgr_custom_empty_factory.cpp should not be the default factory in
as that precludes optimized platform-specific versions.  For macOS this
highlighted a link time issue with component builds, as it competed
with SkFontMgr_mac_ct_factory.cpp.

Begin cleanup by removing that as a common source dependency.  For macOS
this results in the need to explicitly note the Apple freameworks needed
by Skia.  Also remove the duplicate source definition for Android which
is being included for all platforms.

Since the embedder test hashes were generated based upon usage of
SkFontMgr_custom_empty_factory instead of SkFontMgr_mac_ct_factory,
update the sources rules to explicitly note this behavior.  In order to
take advantage of various Skia platform-specific optimizations, these
should be updated in follow-on CLs to update the test hashes as needed
once the font manager is changed.

Bug: pdfium/1830
Change-Id: I63610255bda020d8e25e8f17fd7198deedb5704d