Fix component build when Skia is enabled
SkFontMgr_custom_empty_factory.cpp should not be the default factory as
that precludes use of platform-specific versions.  For macOS this
highlighted a link time issue with component builds, as it competed
with SkFontMgr_mac_ct_factory.cpp.

Begin cleanup by removing SkFontMgr_custom_empty_factory.cpp as a
common source dependency.  At a minimum it is not needed for Android.

For macOS this results in the need to explicitly note the Apple
freameworks needed by Skia.  Also remove the duplicate source definition
for Android factory which has been included for all platforms.

This cleanup also shows that Windows and Linux/ChromeOS builds are
dependent upon the custom empty font manager files.

Once the builds are adjusted it reveals that the embedder test hashes
were generated based upon usage of SkFontMgr_custom_empty_factory.  In
the case of macOS this means that SkFontMgr_mac_ct_factory is not
actually being used.  Similar test hash behavior is also noticed for
Windows, that SkFontMgr_win_dw_factory is not being used to get the
embedder test hashes.

Follow-on CLs should update the build to use the platform-specific
SkFontMgr factories and update the test hashes as needed.

Bug: pdfium:1830
Change-Id: I63610255bda020d8e25e8f17fd7198deedb5704d