Do not invoke V8 CPPCG idler afer heap destroyed
This is a bug in the pdfium_test harness where a V8 cppgc message
loop may be pumped after the heap is destroyed (by the scoped document
going out of scope at the end of ProcessPdf(). We've already idled the
loop to emptyness at line 1136, so there's no need after return.

Bug: chromium:1325664
Change-Id: I34d5a82551789d6359dbb7a58185bb9ff83d6629