Fix a DCHECK() failure inside FPDF_SaveAsCopy().
FPDF_SaveAsCopy() calls CPDF_Name::WriteTo() for name objects, which can
potentially write out an empty name, which is invalid. Add a check for
this and case in CPDF_Name to avoid triggering a DCHECK() failure.

With this fix, the FPDFSaveEmbedderTest.Bug1328389 test case can be
enabld since it no longer crashes. Update the test case's expectations
to reflect what actually happens.

Bug: chromium:1328389
Change-Id: I5fffc93d8811afaa9ba0f0fb501032ebf1165dc8