Fix a DCHECK() failure inside FPDF_SaveAsCopy().
FPDF_SaveAsCopy() calls CPDF_Name::WriteTo() for name objects, but names
can be empty. Adjust WriteTo() to avoid writing out an empty name, as it
is effectively a no-op and the IFX_WriteStream implementation has a
DCHECK() to prevent that.

With this fix, the FPDFSaveEmbedderTest.Bug1328389 test case can be
enabled since it no longer crashes. Update the test case's expectations
to reflect what actually happens.

Bug: chromium:1328389
Change-Id: I5fffc93d8811afaa9ba0f0fb501032ebf1165dc8