Fix a regression in CJBig2_SDDProc::DecodeArith().
In https://pdfium-review.googlesource.com/92890, the refactoring
accidentally changed a uint32_t variable into a uint8_t variable. As a
result, some calculations in CJBig2_SDDProc are incorrect and some
JPEG2000 images fail to render.

Note that there is no test case here, because the PDF that demonstrates
the bug contains PII, and it is not immediately obvious how to create a
JPEG2000 image with the same properties.

Bug: chromium:1329994
Change-Id: I931045d8fb3ad7c565a1d783143dec5f2121c0ae