Fix HeapUnitTest.DISABLED_Bloat and re-enable.
Make FXGCUnitTest use the isolate/platform from the test environment
as set up globally in unit_test_main.cpp, rather than the isolate
created on a per-test basis in FXV8UnitTest. This now pumps the
correct message loop, so that the secheduled GC is executed as
expected, and the test can pass without running out of memory. Note
this affects unit tests only, pdfium_test (and chrome) are not affected.

-- Centralize some code from  GCedTreeNodeUnitTest

Change-Id: I64d654ee1a350352309111d22b0fdcbfbad4d15b