Move CFXJSE_Engine pointer directly into CFXJSE_ResolveProcessor itself.
Currently, this is being passed anew in each call via the
CFXJSE_ResolveProcessor::NodeData.

These are 1:1, and never change, so no need to pass it in via
NodeData on each call. In turn, make one anonymous function
into a method so it can access the engine member.

Next, get CFXJSE_Engine out of the business of creating resolve
processors, so callers can, in theory, pass in a fake one or
something.

Change-Id: I35feab1a69cee89ed9268375c2caef3fa0d9774a