Move the definition of FS_QUADPOINTSF to fpdfview.h.
FS_QUADPOINTSF currently resides in public/fpdf_doc.h, but it is
generally useful in public APIs across multiple headers. Instead of
headers depending on each other in a spaghetti pattern, move
FS_QUADPOINTSF to fpdfview.h, alongside FS_RECTF, FS_SIZEF, and
FS_POINTF.

As part of this CL, remove a now unnecessary include in fpdf_annot.h,
and do IWYU to fix the build for files that depended on this removed
transitive dependency.

Change-Id: I4b9be0053885e8ef2499b7ce7e46b5628274e24b