[Skia] Partially fix FPDFAnnotEmbedderTest.AddAndModifyImage
Make sure to finish painting the background bitmap before applying
the mask layer. This fixes an issue in
FPDFAnnotEmbedderTest.AddAndModifyImage that the dot annotation was
not supposed to be rendered on top of the newly added masked image.

This change appears to fix the issue that only half of the masked
image was rendered by Skia when cache painting is enabled (the
default behavior), which allows us to enable this test. However, the
non-cache painting result still remains unfixed. Therefore, update the
TODO comment to address this remaining issue.

Bug: pdfium:1541,pdfium:1500
Change-Id: Id965812acd31224f7fc650194aa2a33d249294f3