Add a WriteClosedLoop() helper in cpdfsdk_appstream.cpp.
There are several instances of code that generate the same sequence of
drawing commands in cpdfsdk_appstream.cpp. Add a WriteClosedLoop()
function to consolidate these similar instances.

Note that this is not a pure refactor, as some instances that previously
generated stream content of the form "4 5 l f" now generate "4 5 l\nf"
instead.

Change-Id: I06f560146d3c71666371be2f270d44e2c8a6826e