Add a WriteAppendRect() helper in cpdfsdk_appstream.cpp.
There are several instances of code that generate "append rect"
operators in cpdfsdk_appstream.cpp. Add a WriteAppendRect() function to
consolidate these similar instances.

Note that this is not a pure refactor, as some instances that
previously generated stream content of the form "4 5 6 7 re f" now
generate "4 5 6 7 re\nf" instead.

Change-Id: I80e6a4db81f4e0eb01b35285aede4b024d0148c7