Introduce Skia is renderer query
Add a query function to determine if Skia is the renderer in use.  This
is in preparation for enabling Skia and AGG renderers to coexist in a
single build and for the renderer to be selectable at runtime.

This query function enables tests to be incrementally refactored to
always include Skia and AGG definitions in a test build, and to select
at runtime which results to use for expectations.

For now this new query is predictably a direct match of the Skia
defines.  Enabling this to be runtime-driven based on a variable will
come as more changes to the underlying renderer support are filled in.

Bug: pdfium:11
Change-Id: Ie3e74b3ac4355ad90cb478d707356331af52bee7