Rename render create and attach implementations
Update the Skia and AGG specific methods for AttachImpl() and Create()
to have unique names, so that they can coexist in a build.  Introduce
an extra layer which then selects Skia or AGG based upon the current
build definitions.  This is a step towards allowing to choose between
these at runtime.

Bug: pdfium:11
Change-Id: I866d35d599f0525d832cc4adbb775e213e362054