Avoid ambiguity in RetainPtr<> constructors
When two different RetainPtr<> types appear in an expression, and
one is convertible to the other but not vice-versa, SFINAE is not
of any help because the compiler first has to pick a direction in
which to convert (i.e. it won't try both and reject the one that
fails).

-- code taken from base/memory/scoped_refptr.h
-- add a test case that fails to compile without the patch.

Change-Id: I7fea264643bf4b9343336cf68253dcac7199a71b