Convert BinaryBuffer to store a DataVector.
Then BinaryBuffer::DetachBuffer() can avoid a copy and
DetachBufferDeprecated() can be removed. `m_AllocSize` also goes away,
because DataVector can keep track of it.

Converting to only caller to DetachBufferDeprecated() means passing the
data to CPDF_Stream. So CPDF_Stream::TakeData() has to take a DataVector
as well. Though it performs a copy instead of truly taking the data,
because CPDF_Stream does not store a DataVector yet. The pre-existing
version of TakeData() now becomes TakeDataInternal().

Bug: pdfium:1872
Change-Id: I213149cad7ed64bb39795a73110c5a8d15c10073