Add a flag for selecting renderer at runtime
Introduce an experimental config flag to allow for changing the renderer
at runtime.  This flag only takes effect if the config version is 4 or
higher.  When using a config version of 4, it is necessary to specify a
valid renderer type that will be used to specify the renderer type at
runtime.  This is required even if a build does not include Skia
definitions and only has AGG as the available renderer type.

Expose this flag in the pdfium_test sample, via the --use-renderer
command line option.  This option is only available for builds which
have a Skia build definition.

Bug: pdfium:1878
Change-Id: I971f8f7d420015a919316b1b3d57fe3d7bea4ce4