Add a FixedUninitDataVector container.
This is a simple container to hold data of a fixed size. It does not
initialize the underlying memory, so it does not have the same
performance issues as std::vector in debug builds. Its public methods
are very limited and generally requires callers to get to the underlying
data via spans.

Change-Id: Ic995e37707084f940a87a502b25e317d8a71dd5d