[Skia] Remove DCHECKs on RGB channels during upsampling
For a bitmap, if it's 1 bit-per-pixel and it has a color palette, the
palette will contain 2 colors. If the R,G,B channels of either color
have the same values, we can tell those two colors are grayscale and
they will be the only 2 colors used for rendering that bitmap.

However, the R,G,B channel of either color are not always guaranteed
to be equal. In this case, the two colors represents a range of colors
to be used for rendering the bitmap.

This CL removes the incorrect DCHECKs which enforce R,G,B channels to
be equal and adds the detection on when we should use a range of
colors instead of only 2 colors for bitmap rendering.

Meanwhile, rename `color1` and`color2` to `color0` and `color1` to
match the index for CFX_DIBBase::GetPaletteArgb() calls.

Bug: pdfium:1810
Change-Id: Idd83e68a2a654fa5db9c6552e31b42f9b9482acf