Create testing command line helpers
Promote ParseSwitchKeyValue() out of pdfium_test sample into a helper
file so that it can be shared with other testing executables.

Similarly move BuildDefaultRendererType() out of pdfium_test sample.
Rename this to avoid possible confusion with a builder pattern, as
this routine is a query which does not build anything.

Bug: pdfium:1878
Change-Id: I24e4830a5130a1f46f89be68ab9c814aa0b85e9f