Add a FixedZeroedDataVector container.
This is a simple container to hold data of a fixed size. Unlike
FixedUninitDataVector, it does initialize the underlying memory. Since
it always initializes the memory to all zeroes, it does so more
efficiently than std::vector in debug builds. Its public methods are
very limited and generally requires callers to get to the underlying
data via spans.

Change-Id: I4a5e4238ce4833563e157b203df7243379d9121d