Suppress example_062.pdf for Skia and SkiaPaths
When Skia/SkiaPaths is enabled, the 10-point stars have unwanted black
borders compared to the AGG expectations when rendering at the scale
of 1. The unwanted borders will disappear if scale level is larger
enough (e.g when scale = 3).

Suppress this test for Skia and SkiaPaths until the corresponding bug
is fixed.

Bug: pdfium:1896
Change-Id: I07207c899cc22face069e691ea7e227d57786d85