Fix PDF annotation drawing order.
For unknown historical reasons, PDFium draws PDF annotations in reverse
order. This works fine, except when annotations overlap. In the
overlapping case, the annotations that should be drawn on top of other
annotations are actually drawn on the bottom. Flip the drawing order to
be consistent with other PDF software.

Add a simple pixel test to illustrate this issue. Update corpus tests
where changing the drawing order causes small overlaps to be drawn
differently.

Bug: chromium:1304714
Change-Id: I0f66bcb86db817763ed7b870bafea895c15a7962