[Skia] Fix coons patch mesh rendering
When Skia renderer processes coons patch mesh, `paint` has black as the
default color and it does not have a shader type.

According to SkCanvas::drawPatch() documentation [1], when the
interpolated patch `colors` are provided and `paint` does not contain
a shader as the source, `colors` should be used as the destination.

This CL changes the blend mode to SkBlendMode::kDst to match the
documentation.

[1] https://api.skia.org/classSkCanvas.html#a29b048324b32bfc785c0c256d65aa885

Bug: pdfium:1809
Change-Id: I73221c54c9adcb5b77ac276241c9d41341bc39bb