Fix CPDF_DocPageData::MaybePurgeFontFileStreamAcc()
This is never actually purging any stream accessors since the caller
is holding a retained reference and the code is looking to see if
it's map is the only retained reference. Callers must give up their
reference first, and this is accomplished via move semantics.

Code coverage confirms this path was never being hit.

Change-Id: I4a4d801aef94a6611f6d1cfdce4e52a8160454b5