Separate test_runner.py's per-worker state
Separates per-worker process state from the main process's state. This
makes it easier to avoid re-creating objects which should be singletons,
like the SkiaGoldTester.

This change fixes the immediate issue with the SkiaGoldTester, but more
generally, the TestRunner instance shouldn't be pickled to workers; a
future change will fix the remaining issues.

Bug: pdfium:1641
Change-Id: I729a80b05aa7c7d6fa0ad868f4d89e160044342c