Don't pickle TestRunner
Avoids pickling TestRunner by separating state used to run individual
tests from state used to orchestrate the test suite as a whole.

This change has a relatively large delta due to moving code from the
TestRunner class to the _PerProcessState class, but most of the changes
focus on moving properties and initialization code from TestRunner to
_PerProcessConfig or _PerProcessState, along with necessary adjustments
to any call sites in TestRunner.

Bug: pdfium:1641
Change-Id: Ic648f1e72c2bf0f591c2e8c8eff4047b87d0df46