[Skia] Fix issue that a dot annotation is not rendered
An ink dot annotation is considered as a path with 2 identical points
while using Skia or SkiaPaths as the renderer, and SkCanvas::drawPath()
won't draw such paths even if the paint has a
non-zero width.

This CL adds a check whether such a path represents a point and use
SkCanvas::drawPoint() API to render it.

Bug: pdfium:1803
Change-Id: Icfd667be0997abf662d6a04f9621c2bd786eb771