Optimization of decoding large JPX images
We do not use large scales when decoding jpx images:
The JPG image is presented as a set of layers of different scales. With
the cp_reduce parameter, we tell OpenJPEG to ignore <cp_reduce> of the
largest scales. We define the number of layers to ignore as the
logarithm of the ratio the real image size to the requested page size.

This allows us not to waste time to decoding large scales on large
pictures.
We assume that it is possible to ignore all scales that are larger than
the requested rendering size.

Bug: pdfium:1924

Change-Id: I791aa63343f5c32657708003212c8007040e3bc8