[Skia] Fix issue that a dot annotation is not rendered
An ink dot annotation is considered as an unclosed path with 2
identical points. For Skia/SkiaPaths, SkCanvas::drawPath() won't draw
such paths even if the paint has a non-zero width.

This CL adds a check whether such a path represents a point and use
SkCanvas::drawPoint() API to render it.

Bug: pdfium:1803
Change-Id: I1e554a32781288096d7c5f25e2249f1b68ad0a3b