Revert "[Skia] Fix the black borders of small images"
This reverts commit b76ba974234aa48061da6934064cecb4b12731a1.

Reason for revert: Root cause was not the SkMipmapMode type that was used. Changing the SkMipmap happened to fix the issue was because turning on SkMipmapMode::kLinear will activate a newer CPU backend that also makes opaque be enforced if the device is `kOpaque_SkAlphaType`. The correct fix should be changing the renderer device type to `kPremul_SkAlphaType`.

Original change's description:
> [Skia] Fix the black borders of small images
>
> When using SkFilterMode::kLinear to smooth the image rendering result, switch SkMipmapMode to kNearest to avoid the unwanted black borders of small images.
>
> Bug: pdfium:1818,pdfium:1896
> Change-Id: I36edfa53cf21818b9c6ddb95320cab0c71a0f6c6
> Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/100174
> Reviewed-by: Lei Zhang <thestig@chromium.org>
> Commit-Queue: Nigi <nigi@chromium.org>

# Not skipping CQ checks because original CL landed > 1 day ago.

Bug: pdfium:1818,pdfium:1896
Change-Id: I4e26227e79e0730e25ce66a9ff31f60dabec8eae