Check resolution_levels_to_skip value in CJPX_Decoder::Init().
Set a limit for the largest allowed value for the number of resolution
levels to skip for JPEG2000 decoding. This limit is likely unreachable
in production, but can be easily reached by pdf_jpx_fuzzer.

Along the way, remove a redundant check, rename a constant to kFoo, and
change the constant's type to uint8_t.

Bug: chromium:1381880
Change-Id: I6df311af67250bedd42f07d826837ec199716720