Force builds to use C++17 in standalone builds.
The C++ Style Guide [1] currently recommends C++17, but Chromium is
using C++20 for the most part. [2] Change standalone PDFium to use its
own BUILDCONFIG.gn, which overrides the C++ standard config to always
use C++17. This allows PDFium to roll DEPS for //build and get all the
benefits, without surprises to the C++ standard config.

Note that this does not control the build config when PDFium is
embedded in another project.

[1] https://google.github.io/styleguide/cppguide.html#C++_Version
[2] https://crbug.com/1284275

Bug: pdfium:1932
Change-Id: I6b56a392a2b8837504540d48155cb3797afc2df5