Fixed an issue where the installed ComicSansMs font is substituted
During font mapping, in CFX_FolderFontInfo::FindFont(),
the 'Courier New' font is selected if present, in case of Ansi
encoding and Fixed pitch family.

In the Pdf attached to the bug, the `ComicSansMS` font is substituted
by 'Courier New' in the above check, because it has:
- flags value=41, first bit is set == FixedPitch
- WinAisiEncoding,

Since the 'Courier New' check seems like a fallback, I moved it after
inspecting the list of loaded fonts.

Bug: pdfium:1956
Change-Id: Ia3dde2ae932672b8ad22c0e5161c880b4005d547