Drop support for disabling SkiaState cache
Drops alternate code paths for Skia rendering without using the
SkiaState cache. This code path is never used in production, and in many
cases, the cache is required for correct behavior anyway.

Given these differences, disabling the cache is not useful for
debugging, either, and the alternate code paths should be cleaned up.

Bug: pdfium:1845
Change-Id: I74f9ba8d2a4e0d1cd84c5a1bcce0d8b4dd27a7ac