Fixed rendering of symbolic font glyphs used in a composite font
In the PDF attached to the bug, the Wingdings symbol font is used in
a composite font to render a glyph.

The composite font has a ToUnicode map where the referenced symbol
0x39 is mapped to 0xf0fc, the code was ignoring the mapped value and
using the char code 0x39 to look up the Unicode value. Since no
Unicode charmap is defined for this font, we end up with the wrong
charmap and glyph selected.

Bug: pdfium:1946
Change-Id: I40dc68d7256b6fce63cb041ec51e4a0d3c466bbf