[Skia] Fix an assertion error
CFX_SkiaDeviceDriver::SetBitsWithMask() passes 2 bitmap inputs to
Upsample() and Upsample() requires the input bitmaps to be
premultiplied. This requires the CalculateDrawImage() to guarantee
that `bitmap_device1` is qualified to be a premultiplied bitmap
despite its format `FXDIB_Format::kRgb32` doesn't have an alpha
channel.

This CL fill the bitmap's alpha channel with 0xFF so that after
calling CalculateDrawImage(), `bitmap_device1` can be considered
"premultiplied".

Bug: chromium:1395648
Change-Id: I81f32f131be752787d3a1bfde883bf48cd59ac2d