Fix Skia rendering when using `reverse-byte-order` option.
When using `--reverse-byte-order` rendering option, the R and B
channels in the backing SkCanvas are already flipped during the
construction of Skia rendering device. We no longer needed to flip the
byte order in Upsample() calls.

TODO:
Add an embedder test to prevent regression or enable pixel/corpus
tests to run with `reverse-byte-order` option.

Note:
[1] This fix cannot guarantee pixel tests and corpus tests to pass
when using `reverse-byte-order` option.
[2] The fix reduced the pixel tests failure number from 67 to 45 when
using `reverse-byte-order` option. We can probably use an existing
pixel test to create a test (For example: bug_1546.in)

Bug: chromium:1396264, pdfium:1955
Change-Id: I272a70a8e588c772044877058cf9739154b86220